refactor: remove sharp, use UC resize for generate base64 preview and style code
