+++ package.json
@@ -5,7 +5,6 @@
   "dependencies": {
     "@uploadcare/upload-client": "^3.1.1",
     "cheerio": "^1.0.0-rc.10",
-    "gatsby-plugin-sharp": "^4.0.0-next",
     "is-relative-url": "^3.0.0",
     "mdast-util-definitions": "^4.0.0",
     "node-fetch": "^2.6.7",
@@ -17,6 +16,7 @@
   "keywords": [
     "gatsby",
     "gatsby-plugin",
+    "uploadcare",
     "image",
     "markdown",
     "remark",
